Several fixes/enhancements from Martin Becker.
Due to an inadequate size check, input from pipes would be
skipped.  The patch supplied here tests and fixes this.

I have also added a couple of other small improvements:

$^I should only turn on inline-editing if explicitly enabled;
this feature is too dangerous to hide it.

An empty argument list should trigger STDIN processing if
stdin magic is in effect.

It should be documented that inline-editing without backup
files is not supported.

The -I command line option mentioned in the documentation
is actually -i.

The unit test failure rt-56053 seems harmless; that particular
test should be ignored.